+++ tests/utils/businesses/index.js
@@ -1,0 +1,9 @@
+var businessInfo = require('./businessesInfo')
+
+exports.choice = businessInfo.choice;
+exports.utilitySegments = businessInfo.utilitySegments;
+exports.taxSegments = businessInfo.taxSegments;
+exports.businessMap = businessInfo.businessMap;
+exports.randomUtilityBusiness = businessInfo.randomUtilityBusiness;
+exports.randomTaxBusiness = businessInfo.randomTaxBusiness;
+exports.randomBusiness = businessInfo.randomBusiness;