+++ CHANGELOG.md
@@ -55,6 +55,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - DictKey resource to get PIX key's parameters
 - PIX support in Transfer resource
 - BrcodePayment and BrcodePreview support to pay static and dynamic PIX QR Codes
+- TaxPayment object and function calls to allow payment of DAS, DARF and ISS taxes
 
 ## [2.1.0] - 2020-10-28
 ### Added